Assertions

Assertions

Impute.Assertion — Type.

Assertion

An Assertion stores settings for checking the validity of a AbstractArray or Tables.table containing missing values. New assertions are expected to subtype Impute.Assertion and, at minimum, implement the _assert(data::AbstractArray{Union{T, Missing}}, ::<MyAssertion>) method.

source

Impute.assert — Method.

assert(data::AbstractArray, a::Assertion; dims=:)

If the assertion a fails then an ThresholdError is thrown, otherwise the data provided is returned without mutation. See Assertion for the minimum internal _assert call requirements.

Arguments

data::AbstractArray: the data to be impute along dimensions dims

a::Assertion: the assertion to apply

Keywords

dims: The dimension to apply the _assert along (default is :)

Returns

the input data if no error is thrown.

Throws

An error when the test fails

julia> using Test; using Impute: Threshold, ThresholdError, assert

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> @test_throws ThresholdError assert(M, Threshold())
Test Passed
      Thrown: ThresholdError

source

Impute.assert — Method.

assert(table, a::Assertion; cols=nothing)

Applies the assertion a to the table 1 column at a time; if this is not the desired behaviour custom assert methods should overload this method. See Assertion for the minimum internal _assert call requirements.

Arguments

table: the data to impute

a: the assertion to apply

Keyword Arguments

cols: The columns to impute along (default is to impute all columns)

Returns

the input data if no error is thrown.

Throws

An error when any column doesn't pass the test

Example

julia> using DataFrames, Test; using Impute: Threshold, ThresholdError, assert

julia> df = DataFrame(:a => [1.0, 2.0, missing, missing, 5.0], :b => [1.1, 2.2, 3.3, missing, 5.5])
5×2 DataFrame
│ Row │ a        │ b        │
│     │ Float64? │ Float64? │
├─────┼──────────┼──────────┤
│ 1   │ 1.0      │ 1.1      │
│ 2   │ 2.0      │ 2.2      │
│ 3   │ missing  │ 3.3      │
│ 4   │ missing  │ missing  │
│ 5   │ 5.0      │ 5.5      │

julia> @test_throws ThresholdError assert(df, Threshold())
Test Passed
      Thrown: ThresholdError

source

Threshold

Impute.Threshold — Type.

Threshold(; ratio=0.1, weights=nothing)

Assert that the ratio of missing values in the provided dataset does not exceed to specified ratio. If a weights array is provided then the ratio will be calculated as the sum(weights[ismissing.(data)]) / sum(weights)

Keyword Arguments

ratio::Real: Allowed proportion of missing values (should be between 0.0 and 1.0).

weights::AbstractWeights: A set of statistical weights to use when evaluating the importance of each observation. If present a weighted ratio of missing values will be calculated.

source

Impute.ThresholdError — Type.

ThresholdError <: Exception

Is thrown when a Threshold limit is exceed.

Fields

limit::Float64 - the threshold limit.

value::Float64 - the missing data ratio identified

source