Impute.Context — Type.

Context

Stores common summary information for all Imputor types.

Fields

num::Int: number of observations

count::Int: number of missing values found

limit::Float64: allowable limit for missing values to impute

missing::Function: returns a Bool if the value counts as missing.

source

Base.findfirst — Method.

findfirst{T<:Any}(ctx::Context, data::AbstractArray{T, 1}) -> Int

Returns the first not missing index in data.

Arguments

ctx::Context: the context to pass into is_missing

data::AbstractArray{T, 1}: the data array to search

Returns

Int: the first index in data that isn't missing

source

Base.findlast — Method.

findlast{T<:Any}(ctx::Context, data::AbstractArray{T, 1}) -> Int

Returns the last not missing index in data.

Arguments

ctx::Context: the context to pass into is_missing

data::AbstractArray{T, 1}: the data array to search

Returns

Int: the last index in data that isn't missing

source

Base.findnext — Method.

findnext{T<:Any}(ctx::Context, data::AbstractArray{T, 1}) -> Int

Returns the next not missing index in data.

Arguments

ctx::Context: the context to pass into is_missing

data::AbstractArray{T, 1}: the data array to search

Returns

Int: the next index in data that isn't missing

source

Impute.is_missing — Method.

is_missing(ctx::Context, x) -> Bool

Uses ctx.missing to determine if x is missing. If x is a data row or an abstract array then is_missing will return true if ctx.missing returns true for any element. The ctx.count is increased whenever whenever we return true and if ctx.count / ctx.num exceeds our ctx.limit we throw an ImputeError

Arguments

ctx::Context: the contextual information about missing information.

x: the value to check (may be an single values, abstract array or row)

source