Filter

Impute.Filter — Type

Filter([f])

Uses a function f to identify values, rows, columns or slices of data that should be removed during an apply call. The default function f will removing missings, or any rows, columns or slices containing missings.