Impute.impute! — Function.

impute!(imp::Imputor, data::Dataset, limit::Float64=0.1)

Creates a Context using information about data. These include

missing data function:

isnull: if NullableArray or DataTable

isna: if DataArray or DataFrame

isnan: for anything else.

number of elements: *(size(data)...)

Arguments

imp::Imputor: the Imputor method to use

data::Dataset: the data to impute

limit::Float64: missing data ratio limit/threshold (default: 0.1)

Return

Dataset: the input data with values imputed.

source

Impute.impute! — Method.

impute!{T<:Any}(imp::Imputor, ctx::Context, data::Table)

Imputes the data in a DataFrame or DataTable by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

ctx::Context: the contextual information for missing data

data::Table: the data to impute

Returns

Table: the input data with values imputed

source

Impute.impute! — Method.

impute!{T<:Any}(imp::Imputor, ctx::Context, data::AbstractArray{T, 2})

Imputes the data in a matrix by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

ctx::Context: the contextual information for missing data

data::AbstractArray{T, 2}: the data to impute

Returns

AbstractArray{T, 2}: the input data with values imputed

source

Impute.Imputor — Type.

Imputor

An imputor stores information about imputing values in AbstractArrays, DataFrames and DataTables. New imputation methods are expected to sutype Imputor and, at minimum, implement the impute!{T<:Any}(imp::<MyImputor>, ctx::Context, data::AbstractArray{T, 1}) method.

source

Drop

Impute.impute! — Method.

impute!(imp::Drop, ctx::Context, data::Table)

Finds the missing rows in the DataFrame or DataTable and deletes them.

NOTE: this isn't quite as fast as dropnull in DataTables as we're using an arbitrary missing function rather than using the precomputed dt.isnull vector of bools.

Arguments

imp::Drop: this Imputor method

ctx::Context: contextual information for missing data

data::Table: the data to impute

Returns

Table: our data with the missing rows removed.

source

Impute.impute! — Method.

impute!{T<:Any}(imp::Drop, ctx::Context, data::AbstractArray{T, 1})

Uses filter! to remove missing elements from the array.

Arguments

imp::Drop: this Imputor method

ctx::Context: contextual information for missing data

data::AbstractArray{T, 1}: the data to impute

Returns

AbstractArray{T, 1}: our data array with missing elements removed

source

Impute.impute! — Method.

impute!{T<:Any}(imp::Drop, ctx::Context, data::AbstractArray{T, 2})

Finds the missing rows in the matrix and uses a mask (Array{Bool, 1}) to return the data with those rows removed. Unfortunately, the mask approach requires copying the matrix.

NOTES (or premature optimizations):

We use view, but this will change the type of the data by returning a SubArray

We might be able to do something clever by:

reshaping the data to a vector

running deleteat! for the appropriate indices and

reshaping the data back to the desired shape.

Arguments

imp::Drop: this Imputor method

ctx::Context: contextual information for missing data

data::AbstractArray{T, 2}: the data to impute

Returns

AbstractArray{T, 2}: a new matrix with missing rows removed

source

Impute.Drop — Type.

Drop <: Imputor

Removes missing values from the AbstractArray, DataFrame or DataTable provided.

source

Fill

Impute.impute! — Method.

impute!{T<:Any}(imp::Fill, ctx::Context, data::AbstractArray{T, 1})

Computes the fill value if imp.value is a Function (i.e., imp.value(drop(copy(data)))) and replaces all missing values in the data with that value.

source

Impute.Fill — Type.

Fill <: Imputor

Fills in the missing data with a specific value.

Fields

value::Any: A scalar missing value or a function that returns the a scalar if passed the data with missing data removed (e.g, mean)

source

Impute.Fill — Method.

Fill() -> Fill

By default Fill() will use the mean of the existing values as the fill value.

source

Interpolate

Impute.impute! — Method.

impute!{T<:Any}(imp::Interpolate, ctx::Context, data::AbstractArray{T, 1})

Uses linear interpolation between existing elements of a vector to fill in missing data.

WARNING: Missing values at the head or tail of the array cannot be interpolated if there are no existing values on both sides. As a result, this method does not guarantee that all missing values will be imputed.

source

Impute.Interpolate — Type.

Interpolate <: Imputor

Performs linear interpolation between the nearest values in an vector.

source

Last Observation Carried Forward (LOCF)

Impute.impute! — Method.

impute!{T<:Any}(imp::LOCF, ctx::Context, data::AbstractArray{T, 1})

Iterates forwards through the data and fills missing data with the last existing observation.

WARNING: missing elements at the head of the array may not be imputed if there is no existing observation to carry forward. As a result, this method does not guarantee that all missing values will be imputed.

Usage

source

Next Observation Carried Backward (NOCB)

Impute.impute! — Method.

impute!{T<:Any}(imp::NOCB, ctx::Context, data::AbstractArray{T, 1})

Iterates backwards through the data and fills missing data with the next existing observation.

WARNING: missing elements at the tail of the array may not be imputed if there is no existing observation to carry backward. As a result, this method does not guarantee that all missing values will be imputed.

Usage

source

Impute.NOCB — Type.

NOCB <: Imputor

Fills in missing data using the Next Observation Carried Backward (NOCB) approach.

source

Chain

Impute.impute! — Method.

impute!(imp::Chain, missing::Function, data::Dataset; limit::Float64=0.1)

Creates a Context and runs the Imputors on the supplied data.

Arguments

imp::Chain: the chain to run

missing::Function: the missing function to use in the Context to pass to the Imputors

data::Dataset: our data to impute

limit::Float64: the missing data ration limit/threshold

Returns

Dataset: our imputed data

source

Impute.impute! — Method.

impute!(imp::Chain, data::Dataset; limit::Float64=0.1)

Infers the missing data function from the data and passes that to impute!(imp::Chain, missing::Function, data::Dataset; limit::Float64=0.1).

Arguments

imp::Chain: the chain to run

data::Dataset: our data to impute

limit::Float64: the missing data ration limit/threshold

Returns

Dataset: our imputed data

source

Impute.Chain — Type.

Chain <: Imputor

Runs multiple Imputors on the same data in the order they're provided.

Fields

imputors::Array{Imputor}

source

Impute.Chain — Method.

Chain(imputors::Imputor...) -> Chain

Creates a Chain using the Imputors provided (ordering matters).

source