Base.convert — Method.

convert{T<:Any}(::Type{T}, x::Nullable)

Converts the value of a Nullable to the specified type. Needed for casting Nullables to Reals for use in UnitRanges

source

Base.filter! — Method.

filter!(f::Function, a::Union{NullableArray, DataArray})

Allows filtering on NullableArrays and DataArrays, this is pretty much copy-paste from base julia, but they only supports filter! on Array{T, 1} since not all AbstractVectors will implement @deleteat!.

source

RDatasets.dataset — Method.

dataset(m::Module, package_name::AbstractString, Dataset_name::AbstractString)

Function for optional loading of test RDatasets into either DataFrames or DataTables

source