Context

Impute.AbstractContext — Type.

AbstractContext

An imputation context records summary information about missing data for an imputation algorithm. All AbstractContexts are callable with a function, which allows us to write code like:

context() do c
    # My imputation code using a clean context
end

This do-block will pass a fresh context to your code and apply the on_complete function on the resulting data and context state. By default, on_complete will throw an ImputeError if we have too many missing values.

source

Impute.Context — Method.

Context

Records base information about the missing data and assume all observations are equally weighted.

Keyword Arguments

n::Int: number of observations

count::Int: number of missing values found

limit::Float64: portion of total values allowed to be imputed (should be between 0.0 and 1.0).

is_missing::Function: must return a Bool indicating if the value counts as missing

on_complete::Function: a function to run when imputation is complete

source

Impute.WeightedContext — Method.

WeightedContext(wv; limit=1.0, is_missing=ismissing, on_complete=complete)

Records information about the missing data relative to a set of weights. This context type can be useful if some missing observation are more important than others (e.g., more recent observations in time series datasets)

Arguments

wv::AbstractWeights: a set of statistical weights to use when evaluating the importance of each observation. Will be accumulated during imputation.

Keyword Arguments

num::Int: number of observations

s::Float64: sum of the weights of missing values

limit::Float64: portion of total values allowed to be imputed (should be between 0.0 and 1.0).

is_missing::Function: returns a Bool if the value counts as missing

on_complete::Function: a function to run when imputation is complete

source

Base.findfirst — Method.

findfirst(ctx::AbstractContext, data::AbstractVector) -> Int

Returns the first non-missing index in data.

Arguments

ctx::AbstractContext: the context to pass into ismissing!

data::AbstractVector: the data array to search

Returns

Int: the first index in data that isn't missing

source

Base.findlast — Method.

findlast(ctx::AbstractContext, data::AbstractVector) -> Int

Returns the last non-missing index in data.

Arguments

ctx::AbstractContext: the context to pass into ismissing!

data::AbstractVector: the data array to search

Returns

Int: the last index in data that isn't missing

source

Base.findnext — Method.

findnext(ctx::AbstractContext, data::AbstractVector) -> Int

Returns the next non-missing index in data.

Arguments

ctx::AbstractContext: the context to pass into ismissing!

data::AbstractVector: the data array to search

Returns

Int: the next index in data that isn't missing

source

Impute.ismissing! — Method.

ismissing!(ctx::AbstractContext, x) -> Bool

Uses ctx.is_missing to determine if x is missing. If x is a NamedTuple or an AbstractArray then ismissing! will return true if ctx.is_missing returns true for any element. The ctx.count is increased whenever whenever we return true and if ctx.count / ctx.num exceeds our ctx.limit we throw an ImputeError

Arguments

ctx::Context: the contextual information about missing information.

x: the value to check (may be an single values, abstract array or row)

source