Imputors

Imputors

Impute.Imputor — Type.

Imputor

An imputor stores information about imputing values in AbstractArrays and Tables.tables. New imputation methods are expected to subtype Imputor and, at minimum, implement the _impute!(data::AbstractArrays, imp::<MyImputor>) method.

While fallback impute and impute! methods are provided to extend your _impute! methods to n-dimensional arrays and tables, you can always override these methods to change the behaviour as necessary.

source

Impute.impute! — Method.

impute!(data::AbstractArray, imp) -> data

Just returns the data when the array doesn't contain missings

source

Impute.impute! — Method.

impute!(data::AbstractArray{Missing}, imp) -> data

Just return the data when the array only contains missings

source

Impute.impute! — Method.

impute!(data::A, imp; dims=:, kwargs...) -> A

Impute the missing values in the array data using the imputor imp. Optionally, you can specify the dimension to impute along.

Arguments

data::AbstractArray{Union{T, Missing}}: the data to be impute along dimensions dims

imp::Imputor: the Imputor method to use

Keyword Arguments

dims=:: The dimension to impute along. :rows and :cols are also supported for matrices.

Returns

AbstractArray{Union{T, Missing}}: the input data with values imputed

NOTES

Matrices have a deprecated dims=2 special case as dims=: is a breaking change

Mutation isn't guaranteed for all array types, hence we return the result

eachslice is used internally which requires Julia 1.1

Example

julia> using Impute: Interpolate, impute!

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute!(M, Interpolate(); dims=1)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  3.0  4.0  5.0
 1.1  2.2  3.3  4.4  5.5

julia> M
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  3.0  4.0  5.0
 1.1  2.2  3.3  4.4  5.5

source

Impute.impute! — Method.

impute!(table, imp; cols=nothing) -> table

Imputes the data in a table by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

table: the data to impute

Keyword Arguments

cols: The columns to impute along (default is to impute all columns)

Returns

the input data with values imputed

Example

julia> using DataFrames; using Impute: Interpolate, impute

julia> df = DataFrame(:a => [1.0, 2.0, missing, missing, 5.0], :b => [1.1, 2.2, 3.3, missing, 5.5])
5×2 DataFrame
│ Row │ a        │ b        │
│     │ Float64? │ Float64? │
├─────┼──────────┼──────────┤
│ 1   │ 1.0      │ 1.1      │
│ 2   │ 2.0      │ 2.2      │
│ 3   │ missing  │ 3.3      │
│ 4   │ missing  │ missing  │
│ 5   │ 5.0      │ 5.5      │

julia> impute(df, Interpolate())
5×2 DataFrame
│ Row │ a        │ b        │
│     │ Float64? │ Float64? │
├─────┼──────────┼──────────┤
│ 1   │ 1.0      │ 1.1      │
│ 2   │ 2.0      │ 2.2      │
│ 3   │ 3.0      │ 3.3      │
│ 4   │ 4.0      │ 4.4      │
│ 5   │ 5.0      │ 5.5      │

source

Impute.impute! — Method.

impute!(data::T, imp; kwargs...) -> T where T <: AbstractVector{<:NamedTuple}

Special case rowtables which are arrays, but we want to fallback to the tables method.

source

Impute.impute — Method.

impute(data::T, imp; kwargs...) -> T

Returns a new copy of the data with the missing data imputed by the imputor imp. For matrices and tables, data is imputed one variable/column at a time. If this is not the desired behaviour then you should overload this method or specify a different dims value.

Arguments

data: the data to be impute

imp::Imputor: the Imputor method to use

Returns

the input data with values imputed

Example

julia> using Impute: Interpolate, impute

julia> v = [1.0, 2.0, missing, missing, 5.0]
5-element Array{Union{Missing, Float64},1}:
 1.0
 2.0
  missing
  missing
 5.0

julia> impute(v, Interpolate())
5-element Array{Union{Missing, Float64},1}:
 1.0
 2.0
 3.0
 4.0
 5.0

source

Standardize

Impute.Standardize — Type.

Standardize(; values)

Standardize (or replace) various missing data values with missing. This is useful for downstream imputation methods that assume missing data is represented by a missing.

!!! In-place methods are only applicable for datasets which already allowmissing.

Keyword Arguments

value::Tuple: A tuple of values that should be considered missing

Example

julia> using Impute: Standardize, impute

julia> M = [1.0 2.0 -9999.0 NaN 5.0; 1.1 2.2 3.3 0.0 5.5]
2×5 Array{Float64,2}:
 1.0  2.0  -9999.0  NaN    5.0
 1.1  2.2      3.3    0.0  5.5

julia> impute(M, Standardize(; values=(NaN, -9999.0, 0.0)))
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

source

Replace

Impute.Replace — Type.

Replace(; value)

Replace missings with one of the specified constant values, depending on the input type. If multiple values of the same type are provided then the first one will be used. If the input data is of a different type then the no replacement will be performed.

Keyword Arguments

values::Tuple: A scalar or tuple of different values that should be used to replace missings. Typically, one value per type you're considering imputing for.

Example

julia> using Impute: Replace, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, Replace(; values=0.0); dims=2)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  0.0  0.0  5.0
 1.1  2.2  3.3  0.0  5.5

source

Substitute

Impute.Substitute — Type.

Substitute(; statistic=nothing)
Substitute(; robust=true, weights=nothing)

Substitute missing values with a summary statistic over the non-missing values.

Keyword Arguments

statistic: A summary statistic function to be applied to the non-missing values. This function should return a value of the same type as the input data eltype. If this function isn't passed in then the defaultstats function is used to make a best guess.

robust: Whether to use median or mean for continuous datasets in defaultstats

weights: A set of statistical weights to apply to the mean or median in defaultstats.

Default Rules

Our default substitution rules defined in defaultstats are as follows:

mode applies to non-Reals, Bools, and Integers with few unique values.

median is used for all other Real values that aren't restricted by the above rules. Optionally, mean can be used if robust=false. If statistical weights are passed in then a weighted mean/median will be calculated.

Example

julia> using Statistics; using Impute: Substitute, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, Substitute(; statistic=mean ∘ skipmissing); dims=:rows)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  2.66667  2.66667  5.0
 1.1  2.2  3.3      3.025    5.5

source

Simple Random Sample (SRS)

Impute.SRS — Method.

SRS(; rng=Random.GLOBAL_RNG)

Simple Random Sampling (SRS) imputation is a method for imputing both continuous and categorical variables. Furthermore, it completes imputation while preserving the distributional properties of the variables (e.g., mean, standard deviation).

The basic idea is that for a given variable, x, with missing data, we randomly draw from the observed values of x to impute the missing elements. Since the random draws from x for imputation are done in proportion to the frequency distribution of the values in x, the univariate distributional properties are generally not impacted; this is true for both categorical and continuous data.

Keyword Arguments

rng::AbstractRNG: A random number generator to use for observation selection

Example

julia> using Random; using Impute: SRS, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, SRS(; rng=MersenneTwister(1234)); dims=:rows)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  1.0  2.0  5.0
 1.1  2.2  3.3  3.3  5.5

source

Interpolate

Impute.Interpolate — Type.

Interpolate()

Performs linear interpolation between the nearest values in an vector. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

!!! Missing values at the head or tail of the array cannot be interpolated if there are no existing values on both sides. As a result, this method does not guarantee that all missing values will be imputed.

Example

julia> using Impute: Interpolate, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, Interpolate(); dims=:rows)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  3.0  4.0  5.0
 1.1  2.2  3.3  4.4  5.5

source

Last Observation Carried Forward (LOCF)

Impute.LOCF — Type.

LOCF())

Last observation carried forward (LOCF) iterates forwards through the data and fills missing data with the last existing observation. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

See also:

NOCB: Next Observation Carried Backward

!!! Missing elements at the head of the array may not be imputed if there is no existing observation to carry forward. As a result, this method does not guarantee that all missing values will be imputed.

Example

julia> using Impute: LOCF, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, LOCF(); dims=:rows)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  2.0  2.0  5.0
 1.1  2.2  3.3  3.3  5.5

source

Next Observation Carried Backward (NOCB)

Impute.NOCB — Type.

NOCB()

Next observation carried backward (NOCB) iterates backwards through the data and fills missing data with the next existing observation.

See also:

LOCF: Last Observation Carried Forward

!!! Missing elements at the tail of the array may not be imputed if there is no existing observation to carry backward. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

Example

julia> using Impute: NOCB, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, NOCB(); dims=:rows)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  5.0  5.0  5.0
 1.1  2.2  3.3  5.5  5.5

source

K-Nearest Neighbors (KNN)

Impute.KNN — Type.

KNN(; kwargs...)

Imputation using k-Nearest Neighbor algorithm.

Keyword Arguments

k::Int: number of nearest neighbors

dist::MinkowskiMetric: distance metric suppports by NearestNeighbors.jl (Euclidean, Chebyshev, Minkowski and Cityblock)

threshold::AbstractFloat: threshold for missing neighbors

Reference

Troyanskaya, Olga, et al. "Missing value estimation methods for DNA microarrays." Bioinformatics 17.6 (2001): 520-525.

source

SVD

Impute.SVD — Type.

SVD(; kwargs...)

Imputes the missing values in a matrix using an expectation maximization (EM) algorithm over low-rank SVD approximations.

Keyword Arguments

init::Imputor: initialization method for missing values (default: Substitute())

rank::Union{Int, Nothing}: rank of the SVD approximation (default: nothing meaning start and 0 and increase)

tol::Float64: convergence tolerance (default: 1e-10)

maxiter::Int: Maximum number of iterations if convergence is not achieved (default: 100)

limits::Unoin{Tuple{Float64, Float64}, Nothing}: Bound the possible approximation values (default: nothing)

verbose::Bool: Whether to display convergence progress (default: true)

References

Troyanskaya, Olga, et al. "Missing value estimation methods for DNA microarrays." Bioinformatics 17.6 (2001): 520-525.

source