Imputors

Impute.impute! — Method.

impute!(data::AbstractMatrix, imp::Imputor; kwargs...)

Impute the data in a matrix by imputing the values one variable at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

data::AbstractMatrix: the data to impute

imp::Imputor: the Imputor method to use

Keywords

dims: The dimension to impute along (e.g., observations dim)

Returns

AbstractMatrix: the input data with values imputed

Example

julia> using Impute: Interpolate, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, Interpolate(; context=Context(; limit=1.0)); dims=2)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  3.0  4.0  5.0
 1.1  2.2  3.3  4.4  5.5

source

Impute.impute! — Method.

impute!(table, imp::Imputor)

Imputes the data in a table by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

table: the data to impute

Returns

the input data with values imputed

Example

``jldoctest julia> using DataFrames; using Impute: Interpolate, Context, impute julia> df = DataFrame(:a => [1.0, 2.0, missing, missing, 5.0], :b => [1.1, 2.2, 3.3, missing, 5.5]) 5×2 DataFrame │ Row │ a │ b │ │ │ Float64⍰ │ Float64⍰ │ ├─────┼──────────┼──────────┤ │ 1 │ 1.0 │ 1.1 │ │ 2 │ 2.0 │ 2.2 │ │ 3 │ missing │ 3.3 │ │ 4 │ missing │ missing │ │ 5 │ 5.0 │ 5.5 │

julia> impute(df, Interpolate(; context=Context(; limit=1.0))) 5×2 DataFrame │ Row │ a │ b │ │ │ Float64⍰ │ Float64⍰ │ ├─────┼──────────┼──────────┤ │ 1 │ 1.0 │ 1.1 │ │ 2 │ 2.0 │ 2.2 │ │ 3 │ 3.0 │ 3.3 │ │ 4 │ 4.0 │ 4.4 │ │ 5 │ 5.0 │ 5.5 │

source

Impute.impute — Method.

impute(data, imp::Imputor; kwargs...)

Returns a new copy of the data with the missing data imputed by the imputor imp.

Keywords

dims: The dimension to impute along (e.g., observations dim)

source

Impute.Imputor — Type.

Imputor

An imputor stores information about imputing values in AbstractArrays and Tables.tables. New imputation methods are expected to sutype Imputor and, at minimum, implement the impute!(imp::<MyImputor>, data::AbstractVector) method.

source

Impute.splitkwargs — Method.

splitkwargs(::Type{T}, kwargs...) where T <: Imputor -> (imp, rem)

Takes an Imputor type with kwargs and returns the constructed imputor and the unused kwargs which should be passed to the impute! call.

NOTE: This is used by utility methods with construct and imputor and call impute in 1 call.

source

Drop

Impute.DropObs — Type.

DropObs(; context=Context)

Removes missing observations from the AbstractArray or Tables.table provided.

Keyword Arguments

context::AbstractContext=Context(): A context which keeps track of missing data summary information

Example

julia> using Impute: DropObs, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, DropObs(; context=Context(; limit=1.0)); dims=2)
2×3 Array{Union{Missing, Float64},2}:
 1.0  2.0  5.0
 1.1  2.2  5.5

source

Impute.DropVars — Type.

DropVars(; context=Context())

Finds variables with too many missing values in a AbstractMatrix or Tables.table and removes them from the input data.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Examples

julia> using Impute: DropVars, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, DropVars(; context=Context(; limit=0.2)); dims=2)
1×5 Array{Union{Missing, Float64},2}:
 1.1  2.2  3.3  missing  5.5

source

Fill

Impute.Fill — Type.

Fill(; value=mean, context=Context())

Fills in the missing data with a specific value. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

Keyword Arguments

value::Any: A scalar or a function that returns a scalar if passed the data with missing data removed (e.g, mean)

context::AbstractContext: A context which keeps track of missing data summary information

Example

julia> using Impute: Fill, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, Fill(; context=Context(; limit=1.0)); dims=2)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  2.66667  2.66667  5.0
 1.1  2.2  3.3      3.025    5.5

source

Interpolate

Impute.Interpolate — Type.

Interpolate(; context=Context())

Performs linear interpolation between the nearest values in an vector. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

WARNING: Missing values at the head or tail of the array cannot be interpolated if there are no existing values on both sides. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Example

julia> using Impute: Interpolate, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, Interpolate(; context=Context(; limit=1.0)); dims=2)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  3.0  4.0  5.0
 1.1  2.2  3.3  4.4  5.5

source

Last Observation Carried Forward (LOCF)

Impute.LOCF — Type.

LOCF(; context=Context())

Last observation carried forward (LOCF) iterates forwards through the data and fills missing data with the last existing observation. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

See also:

NOCB: Next Observation Carried Backward

WARNING: missing elements at the head of the array may not be imputed if there is no existing observation to carry forward. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Example

julia> using Impute: LOCF, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, LOCF(; context=Context(; limit=1.0)); dims=2)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  2.0  2.0  5.0
 1.1  2.2  3.3  3.3  5.5

source

Next Observation Carried Backward (NOCB)

Impute.NOCB — Type.

NOCB(; context=Context())

Next observation carried backward (NOCB) iterates backwards through the data and fills missing data with the next existing observation.

See also:

LOCF: Last Observation Carried Forward

WARNING: missing elements at the tail of the array may not be imputed if there is no existing observation to carry backward. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Example

julia> using Impute: NOCB, Context, impute

julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0   missing  missing  5.0
 1.1  2.2  3.3       missing  5.5

julia> impute(M, NOCB(; context=Context(; limit=1.0)); dims=2)
2×5 Array{Union{Missing, Float64},2}:
 1.0  2.0  5.0  5.0  5.0
 1.1  2.2  3.3  5.5  5.5

source

Chain

Impute.impute! — Method.

impute!(data, imp::Chain)

Runs the Imputors on the supplied data.

Arguments

imp::Chain: the chain to run

data: our data to impute

Returns

our imputed data

source

Impute.Chain — Type.

Chain <: Imputor

Runs multiple Imputors on the same data in the order they're provided.

Fields

imputors::Array{Imputor}

source

Impute.Chain — Method.

Chain(imputors::Imputor...) -> Chain

Creates a Chain using the Imputors provided (ordering matters).

source

Base.:∘ — Method.

Compose new Imputor chains with the composition operator

Example

julia> using Impute: impute, Interpolate, NOCB, LOCF, Context

julia> ctx = Context(; limit=1.0)
Context(0, 0, 1.0, ismissing, Impute.complete)

julia> imp = Interpolate(; context=ctx) ∘ NOCB(; context=ctx) ∘ LOCF(; context=ctx)
Impute.Chain(Impute.Imputor[Interpolate(2, Context(0, 0, 1.0, ismissing, complete)), NOCB(2, Context(0, 0, 1.0, ismissing, complete)), LOCF(2, Context(0, 0, 1.0, ismissing, complete))])

source